做一点优化

+++ src/main/java/com/huashanlunjian/amara/api/IChartSet.java
@@ -10,7 +10,7 @@ public interface IChartSet {
 
     String getCharter();
 
-    int getBpm();
+    float getBpm();
 
     int getMaxTime();
