+++ src/main/java/com/huashanlunjian/amara/music_game_core/AbstractChart.java
@@ -1,23 +1,21 @@
 package com.huashanlunjian.amara.music_game_core;
 
 import com.huashanlunjian.amara.api.IChartSet;
-import com.huashanlunjian.amara.entity.AbstractNote;
 
 import java.util.Map;
-import java.util.PriorityQueue;
 import java.util.Queue;
 
 public abstract class AbstractChart implements IChartSet {
     protected String title;
     protected String artist;
     protected String charter;
-    protected int bpm = 0;
+    protected float bpm = 0;
     protected int maxTime;
     protected Queue<Map<String, Object>> notes;
     public AbstractChart(Map<String, Object> chart){
 
     }
-    public int getBpm() {
+    public float getBpm() {
         return bpm;
     }
     public String getDifficulty() {
